exceptions were added

+++ src/main/java/org/alex/FileManager/FileManager.java
@@ -8,6 +8,9 @@ public class FileManager {
     public int countFiles(String path) {
         int count = 0;
         File directory = new File(path);
+        if (!directory.exists()) {
+            throw new IllegalStateException("Invalid path! Please recheck the path directory.");
+        }
         if (directory != null) {
             for (File file : directory.listFiles()) {
                 if (file.isFile()) {
@@ -25,6 +28,9 @@ public class FileManager {
     public int countDirs(String path) {
         int count = 0;
         File directory = new File(path);
+        if (!directory.exists()) {
+            throw new IllegalStateException("Invalid path! Please recheck the path directory.");
+        }
         if (directory != null) {
             for (File file : directory.listFiles()) {
                 if (file.isDirectory()) {
@@ -39,6 +45,9 @@ public class FileManager {
 
     public void copy(String from, String to) throws IOException {
         File originalFile = new File(from);
+        if (!originalFile.exists()) {
+            throw new IllegalStateException("No file to copy was found. Please check the path and try again");
+        }
         File newFile = new File(to);
         InputStream origin = new FileInputStream(originalFile);
         OutputStream dest = new FileOutputStream(newFile);
@@ -52,8 +61,18 @@ public class FileManager {
 
     public static void move(String from, String to) {
         File myFile = new File(from);
+        if (!myFile.exists()) {
+            throw new IllegalStateException("No file to move was found. Please check the path and try again");
+        }
         myFile.renameTo(new File(to));
 
+    }
+    public void delete(String path) {
+        File myFile = new File(path);
+        if (!myFile.exists()) {
+            throw new IllegalStateException("No file to delete was found. Please check the path and try again");
+        }
+        myFile.delete();
     }
 
 }
